Renamed folders + Start Recipe button CSS
