+++ src/pages/FoodDetails.js
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
+import './styles/RecipeDetails.css';
 
 import { fetchFoodId } from '../services/fetchFoods';
-import IngredientList from '../components/DetailsPage/IngredientList';
-import DrinkCarousel from '../components/DetailsPage/DrinkCarousel';
+import IngredientList from '../components/RecipeDetails/IngredientList';
+import DrinkCarousel from '../components/RecipeDetails/DrinkCarousel';
 
 function FoodDetails(props) {
   const { id } = props.match.params;
@@ -35,7 +36,7 @@ function FoodDetails(props) {
         <iframe width="420" height="315" src={youtubeURL()} data-testid="video" />
       ) : null}
       <DrinkCarousel />
-      <button data-testid="start-recipe-btn">Iniciar Receita</button>
+      <button data-testid="start-recipe-btn" className="btn-recipe">Iniciar Receita</button>
     </div>
   );
 }
