+++ src/pages/styles/RecipeDetails.css
@@ -1,0 +1,5 @@
+.btn-recipe {
+  bottom: 0;
+  left: 0;
+  position: fixed;
+}
